Add files via upload
This project involves developing a Python-based web scraper that extracts detailed information about the members of the 17th Lok Sabha from a Wikipedia page and stores the data in a structured format. Utilizing `requests` for HTTP requests and `BeautifulSoup` for parsing HTML content, the script navigates through the table of members on the Wikipedia page and extracts key information, including Constituency, Name, Party & a hyperlink to each member's profile. The script efficiently handles variations in the table structure, such as rows with merged cells or vacant entries, ensuring accurate data collection. Once the data is extracted, it is stored in a CSV file, allowing for easy access and further analysis. The CSV file includes headers for the key attributes of each member, providing a structured and organized dataset for future use. This project demonstrates the application of web scraping techniques and data processing for collecting publicly available information and converting it into a usable format.